worksheets/04_dimensionalityReduction/requirements.txt
ONE-api
ibllib
scipy
matplotlib
